+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## Fix
 - [Move /task-forms/{task-id}/variables from rest to enterprise](https://github.com/Alfresco/alfresco-js-api/issues/199)
 - [Added query param for get related contents](https://issues.alfresco.com/jira/browse/ADF-684)
+- [Added query param for create related contents on task and process instance](https://issues.alfresco.com/jira/browse/ADF-930)
 
 ## Features
 - [Complete TypeScript definitions for JS-API](https://github.com/Alfresco/alfresco-js-api/issues/228)